requirements.txt
requests
tqdm==4.31.1
better_bencode==0.2.1
aiopg==0.16.0
aiohttp==3.5.4
aiofiles==0.4.0
cfscrape==2.0.7
chardet==3.0.4
bencode==1.0
SQLAlchemy==1.3.5
aioredis==1.2.0
beautifulsoup4==4.7.1
psycopg2_binary==2.8.3
uvloop==0.12.2
mecab-python3==0.996.2
asyncpg==0.18.3
nose==1.3.7
sanic==19.6.2
ftfy==5.5.1
flask-talisman==0.7.0
geoip2==2.9.0
humanfriendly==4.18
psycopg2==2.8.2
Flask-Caching==1.7.2
Flask-APScheduler==1.11.0
uwsgi==2.0.18
ahocorapy==1.5.2